Hacks which improve throughput and memory usage
These are some hacks to sendfile64_stubs.c to change it so that it does
not use sendfile but instead performs a normal read()/write() copy of
the requested length using a buffer of up to 2MB.

It also contains a hack which forces O_DIRECT on in the output fd. This
should be removed and the code which opens it should use O_DIRECT. Once
this is done there is no longer anything linux-specific in the code.

The buffer needs to turn into something thread safe, and the name should
probably change as it definitely isn't sendfile.

Alternatively the copy could be done entirely in OCaml if there's a way
to do it in large chunks.